split type compatibility helper
